refactor: adopt layered gin project layout
