--- internal/server/router.go
@@ -1,36 +1,0 @@
-package server
-
-import (
-	"fmt"
-	"time"
-
-	"github.com/gin-gonic/gin"
-
-	"smart-mail-relay-go/internal/handlers"
-)
-
-// SetupRouter configures routes and middleware
-func SetupRouter(h *handlers.Handlers) *gin.Engine {
-	gin.SetMode(gin.ReleaseMode)
-	router := gin.New()
-	router.Use(gin.Recovery())
-	router.Use(loggerMiddleware())
-	h.SetupRoutes(router)
-	return router
-}
-
-func loggerMiddleware() gin.HandlerFunc {
-	return gin.LoggerWithFormatter(func(param gin.LogFormatterParams) string {
-		return fmt.Sprintf("%s - [%s] \"%s %s %s %d %s \"%s\" %s\"\n",
-			param.ClientIP,
-			param.TimeStamp.Format(time.RFC1123),
-			param.Method,
-			param.Path,
-			param.Request.Proto,
-			param.StatusCode,
-			param.Latency,
-			param.Request.UserAgent(),
-			param.ErrorMessage,
-		)
-	})
-}
